Update and rename temp.cpp to main.cpp
Changed temp to our main.cpp file. This will be used to run the program. Will be compiled with all other algorithm .cpp's.

Heavily commented so it should be readable. We will dial back and refine pseudocode at a later date for the finished product. Essentially where the //TODO: lines are, we will implement our respective skeleton code and finished algorithm class/function calls as necessary to produce whatever the expected output is. 

The UI is functioning entirely and has an included safety net for  main menu choices out of range. EVENTUALLY; we could add a safety net for when the user tries to insert a string or characters instead of an integer. Not important though.


NOTE: Within mainMenuChoice loop; there is a 'subMenuChoice' variable that is unused. We can use this for any sub menus that may be needed within the various 'if / else if' conditionals relating to the main menu choice. As needed.