Create kochsnowflake.cpp
Current Version for version 2.0; implemented 'recursive koch helper' function which establishes a base case of a parameter indicating 'currentStage' being == to the final stage. This will later be set by a user; but final stage can be manually set to determine which iteration of the koch's snowflake you would like to represent.

For reference: The function takes a start point, end point, current stage as parameters. If the current stage != the final stage; three calculations take place to produce three new coordinates for recursion.

- Next Third Point: Takes the X of the endpoint, subtract the X of the start point, times 0.33 (i.e. a third) then you readd the X of the start point. This establishes a X coordinate that is exactly one third distance away from the previous triangle. Y is calculated in the same manner; but with the Y coordinates.
- Next Two-Third Point: Follows the same formula; but uses 0.66 to move the coordinate exactly two thirds away from the starting position.
- Lastly; next Top point: Uses our find top vector helper; passed the nextTwoThird point as X and oneThird as Y. This will return a coordinate that will represent the tip of the triangle.
With these variables established you can then construct a triangle using these three points; and call the recursive helper again passed the aforementioned points; as well as the start/endpoint to rotate through each side of the overall shape.

*** Other notes; tweaked some variables and reorganized code a bit.