Update kochsnowflake.cpp
Implemented functionality in 'initiate' function to save the window as an image. This is done using two objects; of template class types 'texture' and 'image'. Texture is used to create a storage container for the pixels on the window; which stores them using the .update() member function. Then you can copy the texture object into the image object; where you can then use the .saveToFile() member function to create and output a PNG file.